+++ db/migrate/20150220091326_add_main_pic_to_mokio_menu.rb
@@ -1,0 +1,5 @@
+class AddMainPicToMokioMenu < ActiveRecord::Migration
+  def change
+    add_column :mokio_menus, :main_pic, :string
+  end
+end